@@ -1,59 +1,49 @@
-cmake_minimum_required(VERSION 3.1)
+cmake_minimum_required(VERSION 3.12)
 # Name of the project
-set(PROJECT_NAME LoadCurvesFromMaya)
-project(${PROJECT_NAME})
-#Bring the headers into the project (local ones)
-include_directories(include $ENV{HOME}/NGL/include)
+project(LoadCurvesFromMayaBuild)
+# This is the name of the Exe change this and it will change everywhere
+set(TargetName LoadCurvesFromMaya)
+# Instruct CMake to run moc automatically when needed (Qt projects only)
+set(CMAKE_AUTOMOC ON)
+# find Qt libs
+find_package(Qt5Widgets)
+#find_package(Qt5 COMPONENTS Widgets OpenGL core Gui REQUIRED)
+find_package(fmt CONFIG REQUIRED)
+# use C++ 17
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS ON)
+# Set the name of the executable we want to build
+add_executable(${TargetName})
 
-#the file(GLOB...) allows for wildcard additions of our src dir
-set(SOURCES ${PROJECT_SOURCE_DIR}/src/main.cpp  
+# Add NGL include path
+include_directories(include $ENV{HOME}/NGL/include)
+target_sources(${TargetName} PRIVATE ${PROJECT_SOURCE_DIR}/src/main.cpp  
 			${PROJECT_SOURCE_DIR}/src/NGLScene.cpp  
 			${PROJECT_SOURCE_DIR}/include/NGLScene.h  
 )
-# Set to C++ 11
-set(CMAKE_CXX_STANDARD 11)
+add_compile_definitions(ADDLARGEMODELS)
 
 # see what platform we are on and set platform defines
 if(${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
-	add_definitions(-DDARWIN)
+	message("Mac build")
 	find_library(MACGL OpenGL)
-	set ( PROJECT_LINK_LIBS -lNGL ${MACGL})
-
+	add_definitions(-DGL_SCILENCE_DEPRICATION)
+	set ( PROJECT_LINK_LIBS ${MACGL})
 elseif(${CMAKE_SYSTEM_NAME} MATCHES "Linux")
-	add_definitions(-DLINUX)
-	set ( PROJECT_LINK_LIBS -lNGL -lGL)
-
+	set ( PROJECT_LINK_LIBS -lGL)
 endif()
 
-
-#As were using Qt we need to run moc
-#disable -rdynamic
-SET(CMAKE_SHARED_LIBRARY_LINK_CXX_FLAGS "")
-SET(CMAKE_SHARED_LIBRARY_LINK_C_FLAGS "")
-# define this if you want to include the stanford data sets
-# these are very big and make compilation time huge
-add_definitions(-DADDLARGEMODELS)
-
-# Find includes in corresponding build directories
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
-# Instruct CMake to run moc automatically when needed.
-set(CMAKE_AUTOMOC ON)
-
-add_definitions(-O2 -D_FILE_OFFSET_BITS=64 -fPIC) 
-
-
 # now add NGL specific values
-link_directories( $ENV{HOME}/NGL/lib )
-# as NGL uses Qt we need to define this flag
-add_definitions(-DQT5BUILD)
-# NGL also needs the OpenGL framework from Qt so add it
-find_package(Qt5OpenGL)
-find_package(Qt5Widgets)
-find_package(Qt5Gui)
-find_package(Qt5Core)
+target_link_directories(${TargetName} PRIVATE $ENV{HOME}/NGL/lib )
 
+# add exe and link libs that must be after the other defines
 
-# add exe and link libs this must be after the other defines
-add_executable(${PROJECT_NAME} ${SOURCES})
-target_link_libraries(${PROJECT_NAME} ${PROJECT_LINK_LIBS} Qt5::OpenGL Qt5::Core Qt5::Gui Qt5::Widgets )
+target_link_libraries(${TargetName} PRIVATE ${PROJECT_LINK_LIBS} NGL Qt5::Widgets fmt::fmt-header-only  )
+
+add_custom_target(CopyShaders ALL
+    COMMAND ${CMAKE_COMMAND} -E copy
+    ${CMAKE_CURRENT_SOURCE_DIR}/loads.curve
+    ${CMAKE_CURRENT_BINARY_DIR}/loads.curve
+    ) 
 
